# argus/docs/swagger.yaml
basePath: /
definitions:
...
  argus_app_internal_pkg_response.Result:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  argus_app_internal_service.SaveCameraInput:
    properties:
      name:
        maxLength: 128
        type: string
      remark:
        maxLength: 255
        type: string
      rtspUrl:
        maxLength: 2048
        type: string
      subRtspUrl:
        maxLength: 2048
        type: string
    required:
    - name
    - rtspUrl
    type: object
...
  internal_api.AccessCodesResponse:
    properties:
      code:
        type: integer
      data:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
...
  internal_api.CameraResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/argus_app_internal_model.Camera'
      message:
        type: string
    type: object
...
  internal_api.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  internal_api.LoginResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/argus_app_internal_service.LoginResult'
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: argus 后端 API 接口文档
  title: argus API
  version: "1.0"
paths:
  /api/auth/codes:
    get:
      description: 获取当前登录用户拥有的全部按钮权限码
      produces:
      - application/json
      responses:
        "200":
          description: 权限码列表
          schema:
            $ref: '#/definitions/internal_api.AccessCodesResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 获取当前用户权限码集合
      tags:
      - 认证模块
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录获取 AccessToken，同时在 Cookie 中写入 RefreshToken
      parameters:
      - description: 登录参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回用户信息及 AccessToken
          schema:
            $ref: '#/definitions/internal_api.LoginResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      summary: 用户登录
      tags:
      - 认证模块
...
  /api/auth/refresh:
    post:
      description: 从 HttpOnly Cookie 中读取 RefreshToken，签发新的 AccessToken 并轮换 RefreshToken
        Cookie
      produces:
      - text/plain
      responses:
        "200":
          description: 新的 AccessToken
          schema:
            type: string
        "401":
          description: 未授权或 RefreshToken 无效
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      summary: 刷新令牌
      tags:
      - 认证模块
  /api/camera:
    post:
      consumes:
      - application/json
      description: 新增 RTSP 视频源；保存只做字段与 URL 校验，不强制测活成功
      parameters:
      - description: 创建摄像头参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/argus_app_internal_service.SaveCameraInput'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功的摄像头
          schema:
            $ref: '#/definitions/internal_api.CameraResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 新增摄像头
      tags:
      - 摄像头管理
  /api/camera/{id}:
    delete:
      description: 软删除指定 ID 的摄像头；camera_id 永不复用
      parameters:
      - description: 摄像头ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/internal_api.NilResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 删除摄像头
      tags:
      - 摄像头管理
    put:
      consumes:
      - application/json
      description: 根据 ID 修改摄像头名称、RTSP URL 与备注；配置变更后旧测活结果视为不适用于当前配置
      parameters:
      - description: 摄像头ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新摄像头参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/argus_app_internal_service.SaveCameraInput'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的摄像头
          schema:
            $ref: '#/definitions/internal_api.CameraResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 更新摄像头
      tags:
      - 摄像头管理
...
  /api/camera/batch:
    delete:
      consumes:
      - application/json
      description: 批量软删除指定 ID 列表的摄像头
      parameters:
      - description: 批量删除参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.BatchDeleteCameraInput'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/internal_api.NilResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 批量删除摄像头
      tags:
      - 摄像头管理
  /api/camera/page:
    get:
      description: 分页查询摄像头数据，支持名称模糊筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页条数
        in: query
        name: pageSize
        type: integer
      - description: 名称模糊查询
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 摄像头分页数据
          schema:
            $ref: '#/definitions/internal_api.CameraPageResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 分页获取摄像头列表
      tags:
      - 摄像头管理
  /api/camera/probe:
    post:
      consumes:
      - application/json
      - application/json
      description: |-
        对 RTSP 配置执行测活（TCP 优先，失败回退 UDP，首帧即成功）。测活失败也返回 code=0，结果在 data.status 中
        对 RTSP 配置执行测活（TCP 优先，失败回退 UDP，首帧即成功）。测活失败也返回 code=0，结果在 data.status 中
      parameters:
      - description: 测活参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.ProbeCameraInput'
      - description: 测活参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.ProbeCameraInput'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: 测活结构化结果
          schema:
            $ref: '#/definitions/internal_api.ProbeResultResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: 摄像头测活
      tags:
      - 摄像头管理
      - 摄像头管理
  /api/dept:
    post:
      consumes:
      - application/json
      description: 新增部门节点
      parameters:
      - description: 创建部门参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/argus_app_internal_service.SaveDeptInput'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功的部门信息
          schema:
            $ref: '#/definitions/internal_api.DeptResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 创建部门
      tags:
      - 部门模块
  /api/dept/{id}:
    delete:
      description: 删除指定 ID 的部门（有子部门或包含用户时不可删除）
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/internal_api.NilResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 删除部门
      tags:
      - 部门模块
    put:
      consumes:
      - application/json
      description: 更新指定 ID 的部门信息
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新部门参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/argus_app_internal_service.SaveDeptInput'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的部门信息
          schema:
            $ref: '#/definitions/internal_api.DeptResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 更新部门
      tags:
      - 部门模块
  /api/dept/tree:
    get:
      description: 查询系统所有部门并构建为树形结构
      produces:
      - application/json
      responses:
        "200":
          description: 部门树数据
          schema:
            $ref: '#/definitions/internal_api.DeptTreeResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 获取全量部门树
      tags:
      - 部门模块
  /api/file/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传图片或 PDF 文件，返回不需要认证即可访问的文件 URL
      parameters:
      - description: 上传文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            $ref: '#/definitions/internal_api.FileUploadResponse'
        "400":
          description: 文件参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "500":
          description: 存储失败
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 上传文件
      tags:
      - 文件模块
  /api/menu:
    post:
      consumes:
      - application/json
      description: 新增菜单、目录或权限按钮节点
      parameters:
      - description: 创建菜单参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/argus_app_internal_service.SaveMenuInput'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功的菜单信息
          schema:
            $ref: '#/definitions/internal_api.MenuResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 创建菜单/按钮
      tags:
      - 菜单模块
...
  /api/menu/all:
    get:
      description: 根据当前用户角色权限，生成用于前端动态挂载路由的菜单结构
      produces:
      - application/json
      responses:
        "200":
          description: 前端动态路由树
          schema:
            $ref: '#/definitions/internal_api.UserMenuTreeResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 获取当前用户可访问菜单树
      tags:
      - 菜单模块
  /api/menu/tree:
    get:
      description: 查询系统所有菜单/按钮并构建为树形结构（用于菜单管理页面）
      produces:
      - application/json
      responses:
        "200":
          description: 菜单树数据
          schema:
            $ref: '#/definitions/internal_api.MenuTreeResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 获取全量菜单树
      tags:
      - 菜单模块
...
securityDefinitions:
  BearerAuth:
    description: 在 Header 中传入 `Bearer <token>` 进行身份认证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"